feat/added apidoc for swagger; refactor verification controller
